Simplifying markup, css and design
jwforres asked me to see if we could simplify the design a bit now
that many of the planned enhancements from ajacobs21e’s were cut from
the implementation.  I initially set out to just mock the page so I
could make a screenshot, but ended up refactoring in order to figure
out options for a more condensed header.  MarkDeMaria, I’d like to
walk you through the changes when you’re in the office.